Fix BF16 numpy conversion error in get_probs
BF16 tensors cannot be directly converted to numpy arrays.
Cast to float32 before .cpu().numpy().

+++ colab_deberta_kfold.py
@@ -203,7 +203,7 @@ def get_probs(model, loader, desc=""):
         with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=USE_BF16):
             logits = model(input_ids=ids, attention_mask=mask).logits
         probs = torch.softmax(logits, dim=-1)
-        all_probs.append(probs.cpu().numpy())
+        all_probs.append(probs.float().cpu().numpy())
     return np.vstack(all_probs)
